Update DistribPassCrackAttempt.yaml
The rule isn’t firing as expected when we tested it against the configured thresholds in the rule. After some digging in the rule logic I came to a conclusion that the rule logic is flowed and will never fire as due to the “summarize” statement highlighted below the location count will always be equal to 1.